Add files via upload
Arduino Uno board
4 LEDs
4 220 ohms resistors
5 male-to-male jumper wires
Breadboard
The project involves counting from 0 to 15 in binary and displaying the count on the LEDs. The code uses a while loop to count and the bitRead() function to turn on or off the appropriate LEDs based on the binary representation of the count. The circuit diagram is included in the repository to show how to connect the LEDs and resistors to the Arduino Uno board using a breadboard.